Refactor to PostActivity
