+++ presentation/src/main/java/com/pocs/presentation/view/post/detail/PostDetailViewModel.kt
@@ -1,8 +1,8 @@
-package com.pocs.presentation.view.detail.article
+package com.pocs.presentation.view.post.detail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.pocs.presentation.mock.mockArticleDetailItems
+import com.pocs.presentation.mock.mockPostDetailItems
 import com.pocs.presentation.model.ArticleDetailUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -12,14 +12,14 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class ArticleDetailViewModel @Inject constructor() : ViewModel() {
+class PostDetailViewModel @Inject constructor() : ViewModel() {
 
     private val _uiState = MutableStateFlow(ArticleDetailUiState())
     val uiState = _uiState.asStateFlow()
 
     fun loadArticle(id: Int) {
         viewModelScope.launch {
-            val article = mockArticleDetailItems.first { it.id == id }
+            val article = mockPostDetailItems.first { it.id == id }
             _uiState.update { it.copy(article = article) }
         }
     }
